Update use of trading manager.
The BaseTradingStrategy class' default
constructor initialises a TradingManager.
So now this can be hidden from the students
preventing them from giving them a bigger
initial value.

We have also created a new Factory that controls
when the trading strategy gets created so they
do not need to manage the life cycle of their
own variables.